Modify 세트 수정 텍스트 제한 및 수정 프로세스 수정

+++ lib/exercise/model/setInfo_model.dart
@@ -11,7 +11,7 @@ class SetInfo {
   @HiveField(2)
   int? reps;
   @HiveField(3)
-  int? weight;
+  double? weight;
   @HiveField(4)
   int? seconds;
 
@@ -25,7 +25,7 @@ class SetInfo {
   SetInfo copyWith({
     int? index,
     int? reps,
-    int? weight,
+    double? weight,
     int? seconds,
   }) =>
       SetInfo(
